IOEQ: removed, fixes #8866 TaskQueue: queue for some longer taking tasks SecondLevelLauncher: do not remove/commit the line! DataBaseConnector: -modified so we can use it for importing of old database -removed obsolete sync Modified every ScheduledExecutorService to use DelayedRunnable.getNewScheduledExecutorService() AccountController: -modified list to only return accounts with existing plugins (outdated accounts will not be returned) -modified updateAccountInfo to proper use PluginClassLoaderChild DialogHandler: updated to use Challenge.getDomainInfo GeneralSettings: introduced getMaxForcedDownloads PluginFinder: fixed to check for plugin.isPremium LazyHostPlugin: removed getIcon SkipReason: added FILE_IN_PROCESS, PLUGIN_DEFECT, moved translation into enum LogController: fixed for new Queue stuff FilterTable: -use one static DelayedRunnable to avoid heavy cpu usage -removed init QuickFilterTables: fixed init method LinkGrabberOverview/DownloadTableOverview: -optimized to use less cpu -can be gc again (when closed) -avoid cpu usage when not visible (tab change) ContextMenu: -changed all actions to use the correct queue -actions that use ValidateArchiveAction use their own thread, to avoid blocking -fixed many actions to work again even without any current selection LinkGrabberPanel/DownloadsPanel: overview can be gc now DataEntry: fixed, was using hard reference to DownloadsTableModel Columns: updated to use new downloadLink methods PackageControllerTableModel: -changed repaint/recreate to not work inside EDT -toggleFilePackageExpand: fixed to use current selection and not delayed one PackageController: -fixed to use current selection and not delayed one DownloadFolderChooserDialog: -updated handleNonExistingFolders to return folder creation state -moved isDownloadFolderValid to DownloadWatchDog MainToolbar/Menu/Manager: use own delayer SearchFields: use own delayer Extraction: -added synchronized addPassword -write crashlog to log -added Seven7ExtractCallback for 7z workaround extraction (avoids jvm crash for some 7z files) can be used for other archives as well -updated MultiCallback to return written bytes -extracted setLastModifiedDate into own method (because it is used at different places) -modified extract to use new Seven7ExtractCallback for 7z files -extracted getExtractFilePath into own method (because it is used at different places) -modified addToQueue to force ask for unknown password when archive is manually extracted (else default value from advanced settings will be used) ExtractionConfigPanel: -moved updateContents into queue because jsonstorage can take a while for pw list and other stuff ArchiveValidator: fixed sync issue DownloadLinkArchiveFactory: fixed getGuessedPasswordList IEExctraction: added getLogger ExtractionStatus: moved translation into enum ArchiveSettings/ArchiveController: -fixed JSonStorage.toString to proper JSonStorage.serializeToJson -fixed hashmap sync issue PackagizerController/LinkFilterController: -fixed some sync issues, never throw update when holding locks ContextMenuManager: fixed JSonStorage.toString to proper JSonStorage.serializeToJson Challenge: -added getDomainInfo PollingAPI: fixed sync issue ExtractionAPI: changed to use new addPassword method RemoteAPIController: fixed possible npe OldRAFDownloadSystem: -fixed proper open/close of connections -fixed proper abort -fixed possible npe because of wrong init order in OLDRAFDownload constructor -fixed error, no longer can block/be executed multiple times -use downloadWatchDog for local file checks -updated terminate for faster close -proper locking of part/complete files PluginException: can be used now with SkipReason as well Plugin: used captchas cleanup moved to cleanup() LinkStatus: removed many outdated linkstati FilePackage: -updated defaultFilePackage -added hasNotificationListener to avoid object creation when no listener is attached DownloadLink: -fixed readObject, transient variables need to be initialized here -added hasNotificationListener to avoid object creation when no listener is attached -added getFileOutput(boolean ignoreUnsafe, boolean ignoreCustom) -fixed getName to avoid npe -updated reset -many states are now atomic -updated setName functions to use less cpu -added getPreviousParentNodeID to get uniqueID from previous package (can be used to know source/dest when link got moved) -getDomainInfo now uses plugin for domaininfo creation -_setFilePackage now saved last FilePackage destination for proper deletion of link BuyAction: use proper selection MyJDownloaderSettingsPanel: -move connect/disconnect into queue, not in EDT! -pending is same as connected, to avoid confusion OpenDefaultDownloadFolderAction: fixed, now use correct method MainTabbedPane: added notifyCurrentTab JDGUI: moved isCurrentPanel into EDT PackageController: -each controller has its own queue now AbstractPackageChildrenNode: now provides getPreviousParentNodeID AbstractNodeNotifier: now needs hasNotificationListener LinkCrawlerThread: set min thread priority CrawledPackage/CrawledLink: -added hasNotificationListener to avoid object creation when no listener is attached LinkCollector: -changed saving -updated delayer to save when possible -moved clear into queue -avoid File.listFiles, now saves result -moved init into queue to avoid blocking of maininit -added debug to show packages/links that are loaded -added debug for loading, when error occurs -backup broken zip files for later checking/restoring LinkChecker:set min thread priority FavIcons: now try out all subdomains till first valid favicon is found DownloadController: -onShutDown: save first, then stop downloadwatchdog, save again. so we can save even when downloadwatchdog stop takes longer/too long -use 2 delayer now, one for changes, one for downloads -modified addAllAt to remember result of folder creation -removed getFirstLinkThatBlocks -moved init into queue to avoid blocking of maininit -added debug to show packages/links that are loaded -added debug for loading, when error occurs -backup broken zip files for later checking/restoring DownloadWatchDog: -major refactoring -now provides a queue to run tasks synch with DownloadWatchDog (reset, delete, resume...) -moved all session relevant information into DownloadSession Object -added resume/delete/reset functions to work at correct sync moment -rewrote start/stop of SingleDownloadControllers -fixed proper use of forceDownloads -added validateDestination -new SingleDownloadControllerActivator to provide all information for next Download -added nextActivator to provide next possible SingleDownloadControllerActivator -delete now locks the file before real delete, to avoid sync issues FileAccessManager: -updated lock to work again when called by same lockholder -added unlockAllHeldby ,isLocked SingleDownloadController: -updated abort -sync check for "already working" -added support for SkipReason PluginExceptions
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@22160 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/org/jdownloader/utils/JDFileUtils.java
@@ -8,33 +8,41 @@ import org.appwork.utils.Files;
 import org.appwork.utils.os.CrossSystem;
 
 public class JDFileUtils extends Files {
+
+    private static boolean supported = false;
+    static {
+        try {
+            if (CrossSystem.isWindows()) {
+                supported = com.sun.jna.platform.win32.W32FileUtils.getInstance().hasTrash();
+            } else if (CrossSystem.isMac()) {
+                supported = com.sun.jna.platform.mac.MacFileUtils.getInstance().hasTrash();
+            }
+        } catch (final Throwable e) {
+            e.printStackTrace();
+        }
+    }
+
     /**
      * @param f
      */
 
     public static boolean isTrashSupported() {
-        if (CrossSystem.isWindows()) {
-            return com.sun.jna.platform.win32.W32FileUtils.getInstance().hasTrash();
-        } else if (CrossSystem.isMac()) { return com.sun.jna.platform.mac.MacFileUtils.getInstance().hasTrash(); }
-
-        return false;
+        return supported;
     }
 
     public static void moveToTrash(File... files) throws IOException {
-        // .moveToTrash()
-
-        if (CrossSystem.isWindows()) {
-            for (File f : files) {
-                if (!f.exists()) { throw new FileNotFoundException(f.getAbsolutePath()); }
+        if (isTrashSupported()) {
+            if (CrossSystem.isWindows()) {
+                for (File f : files) {
+                    if (!f.exists()) { throw new FileNotFoundException(f.getAbsolutePath()); }
+                }
+                com.sun.jna.platform.win32.W32FileUtils.getInstance().moveToTrash(files);
+            } else if (CrossSystem.isMac()) {
+                for (File f : files) {
+                    if (!f.exists()) { throw new FileNotFoundException(f.getAbsolutePath()); }
+                }
+                com.sun.jna.platform.mac.MacFileUtils.getInstance().moveToTrash(files);
             }
-            com.sun.jna.platform.win32.W32FileUtils.getInstance().moveToTrash(files);
-        } else if (CrossSystem.isMac()) {
-            for (File f : files) {
-                if (!f.exists()) { throw new FileNotFoundException(f.getAbsolutePath()); }
-            }
-            com.sun.jna.platform.mac.MacFileUtils.getInstance().moveToTrash(files);
-
         }
-
     }
 }
